% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{essentialGenes}
\alias{essentialGenes}
\title{List of essential gene symbols}
\format{
List of essential gene symbols
}
\usage{
data(essentialGenes, package="crisprBase")
}
\description{
List of essential gene symbols
}
\details{
List of essential gene symbols
}
\keyword{datasets}
