% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuideSetToBed.R
\name{GuideSetToBed}
\alias{GuideSetToBed}
\title{Save GuideSet object to a BED file}
\usage{
GuideSetToBed(
  guideSet,
  filename = "output.bed",
  overwrite = FALSE,
  trackName = "My gRNA track",
  trackDescription = "Something about my track",
  browserChr = "chr3",
  browserStart = 10141008,
  browserEnd = 10152220,
  browserVisibility = 3,
  scoreColumn = NULL,
  style = c("ucsc", "epiviz")
)
}
\arguments{
\item{guideSet}{A \code{GuideSet} object from crisprDesign.}

\item{filename}{String specifying filename of the BED file.}

\item{overwrite}{Should the file be overwritten if it exists?
FALSE by default.}

\item{trackName}{String specifying the name of the genomic track.}

\item{trackDescription}{String specifying description of the genomic track.}

\item{browserChr}{String specifying default chromosome location of the
genomic track. "Chr3" by default.}

\item{browserStart}{Integer specifying default chromosome start of the 
genomic track.}

\item{browserEnd}{Integer specifying default chromosome end of the 
genomic track.}

\item{browserVisibility}{Integer specifying the default genomic track
visibility.}

\item{scoreColumn}{String specifying column name in \code{mcols(guideSet)}
to be used as the score for the BED file. NULL by default.}

\item{style}{Which bed format style should be used?
For UCSC genome browser, specify "ucsc" (default).
For Epiviz, specify "epiviz".}
}
\value{
No object returned. Object is saved into a BED file as a side 
    effect.
}
\description{
Save GuideSet object to a BED file
}
\examples{
library(crisprDesign)
data(guideSetExampleFullAnnotation, package="crisprDesign")
GuideSetToBed(guideSetExampleFullAnnotation,
              filename=tempfile())

}
\author{
Jean-Philippe Fortin
}
