% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonEssentialGenes}
\alias{nonEssentialGenes}
\title{List of non-essential gene symbols}
\format{
List of non-essential gene symbols
}
\usage{
data(nonEssentialGenes, package="crisprBase")
}
\description{
List of non-essential gene symbols
}
\details{
List of non-essential gene symbols
}
\keyword{datasets}
