% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqliteToGuideSet.R
\name{SqliteToGuideSet}
\alias{SqliteToGuideSet}
\title{Retrieve a GuideSet object from a SQLite database}
\usage{
SqliteToGuideSet(
  dbfile,
  crisprNuclease = NULL,
  bsgenome = NULL,
  targetOrigin = c("bsgenome", "customSequences"),
  customSequences = NULL,
  id.field = "ID",
  ...
)
}
\arguments{
\item{dbfile}{String specifying filename of the database.}

\item{crisprNuclease}{\code{CrisprNuclease} object from the 
crisprBase package representing the nuclease used to 
build the database.}

\item{bsgenome}{\code{BSgenome} object representing the genome
used to build the database.}

\item{targetOrigin}{String specifying how the GuideSet was 
constructed. Must be either "bsgenome" or "customSequences".}

\item{customSequences}{Optional DNA sequences used to build the 
underlying GuideSet objects. Must only be provided when 
targetOrigin is set to "customSequences".}

\item{id.field}{String specifying the name of the column
used to store gRNA ids.}

\item{...}{Other arguments to pass to GuideSet constructor function.}
}
\description{
Save GuideSet object as a SQLite database.
}
\examples{
library(crisprDesign)
data(guideSetExampleFullAnnotation, package="crisprDesign")
dbfile <- tempfile()
GuideSetToSqlite(guideSetExampleFullAnnotation, dbfile)

library(crisprBase)
library(BSgenome.Hsapiens.UCSC.hg38)
data(SpCas9)
crisprNuclease <- SpCas9
bsgenome <- BSgenome.Hsapiens.UCSC.hg38
gs <- SqliteToGuideSet(dbfile,
                       bsgenome=bsgenome,
                       crisprNuclease=crisprNuclease,
                       seqinfo=seqinfo(bsgenome),
                       seqlengths=seqlengths(bsgenome))

}
\author{
Jean-Philippe Fortin
}
